patched various bugs in standard impl.

+++ src/main/java/com/g7/CPEN431/A7/consistentMap/ConsistentMap.java
@@ -297,7 +297,7 @@ public class ConsistentMap {
      * @return whether the server exist in the ring
      */
     public boolean hasServer(InetAddress addr, int port){
-        long hashcode = new VNode(new ServerRecord(addr, port), 0).getHash();
+        int hashcode = new VNode(new ServerRecord(addr, port), 0).getHash();
         lock.readLock().lock();
         boolean hasKey = ring.containsKey(hashcode);
         lock.readLock().unlock();
